// src/types/types.d.ts
export type Pokemon = {
    name: string;
    id: string;
    imgSrc: string;
}

export type PokemonDetails = {
    name: string;
    id: string;
    imgSrc: string;
    hp: number;
    attack: number;
    defense: number;
    special_attack: number;
    special_defense: number;
    speed: number;
}
    